requirements.txt
django>=1.4.8
djangorestframework